SOLR-6159: A ZooKeeper session expiry during setup can keep LeaderElector from joining elections
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1606997 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/cloud/ElectionContext.java
@@ -51,7 +51,7 @@ public abstract class ElectionContext {
   final ZkNodeProps leaderProps;
   final String id;
   final String leaderPath;
-  String leaderSeqPath;
+  volatile String leaderSeqPath;
   private SolrZkClient zkClient;
 
   public ElectionContext(final String coreNodeName,
@@ -66,12 +66,16 @@ public abstract class ElectionContext {
   public void close() {}
   
   public void cancelElection() throws InterruptedException, KeeperException {
-    try {
-      log.info("canceling election {}",leaderSeqPath );
-      zkClient.delete(leaderSeqPath, -1, true);
-    } catch (NoNodeException e) {
-      // fine
-      log.warn("cancelElection did not find election node to remove {}" ,leaderSeqPath);
+    if( leaderSeqPath != null ){
+      try {
+        log.info("canceling election {}",leaderSeqPath );
+        zkClient.delete(leaderSeqPath, -1, true);
+      } catch (NoNodeException e) {
+        // fine
+        log.warn("cancelElection did not find election node to remove {}" ,leaderSeqPath);
+      }
+    } else {
+      log.warn("cancelElection skipped as this context has not been initialized");
     }
   }
